+++ EventAnalyzer/src/EventVariableOSSF.cpp
@@ -28,7 +28,7 @@ bool EventVariableOSSF::calculate(BaseHandler* handler) {
 			if(!m_includeFakes && v[j]->getVariable("fakeRole", fakeRole) && fakeRole > 0) {
 				continue;
 			}
-			double charge1 = 0;
+			int charge1 = 0;
 			bool charge1set = (v[j])->getVariable("CHARGE",charge1);
 			if(!charge1set) {
 				continue;
@@ -38,7 +38,7 @@ bool EventVariableOSSF::calculate(BaseHandler* handler) {
 				if(!m_includeFakes && v[k]->getVariable("fakeRole", fakeRole) && fakeRole > 0) {
 					continue;
 				}
-				double charge2 = 0;
+				int charge2 = 0;
 				bool charge2set = (v[k])->getVariable("CHARGE",charge2);
 				if(!charge2set) {
 					continue;
